requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.3
greenlet==3.2.3
h11==0.16.0
idna==3.10
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.2.3
playwright==1.53.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
